docker_stack/AppFiles/requirements.txt
Django==2.1.7
django-filter==2.1.0
django-taggit==0.24.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.2
drf-writable-nested==0.5.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
six==1.12.0
gunicorn==19.9.0
django-cors-headers==3.0.2
